+++ app/infrastructure/services/__init__.py
@@ -9,15 +9,12 @@ from infrastructure.services.dependencies import (
     SettingsDep,
     MaxMindClientDep,
     EventDispatcherDep,
-    TranslationServiceDep,
 )
 from infrastructure.services.providers import (
     get_app_settings,
     get_settings,
     get_maxmind_client,
     get_event_dispatcher,
-    get_translation_service,
-    t,
 )
 
 __all__ = [
@@ -26,13 +23,9 @@ __all__ = [
     "SettingsDep",
     "MaxMindClientDep",
     "EventDispatcherDep",
-    "TranslationServiceDep",
     # Core providers
     "get_app_settings",
     "get_settings",
     "get_maxmind_client",
     "get_event_dispatcher",
-    "get_translation_service",
-    # Translation helper
-    "t",
 ]
