+++ app/infrastructure/services/dependencies.py
@@ -11,13 +11,11 @@ from fastapi import Depends
 from infrastructure.clients.maxmind import MaxMindClient
 from infrastructure.configuration import AppSettings, Settings
 from infrastructure.events.service import EventDispatcher
-from infrastructure.i18n.service import TranslationService
 from infrastructure.services.providers import (
     get_app_settings,
     get_event_dispatcher,
     get_maxmind_client,
     get_settings,
-    get_translation_service,
 )
 
 # Settings dependency
@@ -33,13 +31,10 @@ MaxMindClientDep = Annotated[MaxMindClient, Depends(get_maxmind_client)]
 # Events dispatcher dependency
 EventDispatcherDep = Annotated[EventDispatcher, Depends(get_event_dispatcher)]
 
-# Translation service dependency
-TranslationServiceDep = Annotated[TranslationService, Depends(get_translation_service)]
-
 
 __all__ = [
     "SettingsDep",
     "AppSettingsDep",
+    "MaxMindClientDep",
     "EventDispatcherDep",
-    "TranslationServiceDep",
 ]
