+++ app/packages/access/sync/interactions/slack.py
@@ -13,10 +13,10 @@ from typing import TYPE_CHECKING, Any, Dict
 
 import structlog
 
+from infrastructure.i18n import t
 from infrastructure.idempotency import get_idempotency_service
 from infrastructure.platforms.models import CommandPayload, CommandResponse
 from infrastructure.platforms.parsing import Argument, ArgumentType
-from infrastructure.services import t
 from packages.access.sync.interactions.ingress import (
     enqueue_platform_sync,
     enqueue_user_sync,
