+++ app/server/lifespan.py
@@ -13,7 +13,7 @@ from infrastructure.configuration.infrastructure.server import (
     get_server_settings,
 )
 from infrastructure.directory import get_directory_provider
-from infrastructure.i18n.resources import I18nResourceSpec
+from infrastructure.i18n import I18nResourceSpec, get_translation_service
 from infrastructure.logging.setup import configure_logging
 from infrastructure.platforms import get_platform_service
 from infrastructure.plugins import (
@@ -23,7 +23,6 @@ from infrastructure.plugins import (
 from infrastructure.security import get_jwks_manager
 from infrastructure.services import (
     get_settings,
-    get_translation_service,
 )
 from jobs import scheduled_tasks
 from modules import (
